Aus Sammlung entfernen + Labelseite
Auf der Versions-Seite kann man jetzt immer schön im Wechsel eine
Version in seine Sammlung aufnehmen und wieder daraus löschen.

Außerdem ist mit der Fertigstellung der Labelseite die letzte Dummy-
Seite aus unserer Anwendung verschwunden.

Nach einem Test durch euch würde ich den aktuellen Stand als Version
1.0!!! in den Master-Branch mergen.

+++ src/main/java/de/yellow/medienverwaltung/api/ReleaseController.java
@@ -11,41 +11,41 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import de.yellow.medienverwaltung.business.MasterService;
 import de.yellow.medienverwaltung.business.ReleaseService;
 import de.yellow.medienverwaltung.database.entity.Release;
 
 @Controller
 public class ReleaseController extends AbstractController {
 
-    @RequestMapping(value = "api/releases/{id}/", method = RequestMethod.GET, produces = "application/json")
-    public ResponseEntity<Release> getRelease(@PathVariable("id") int id) {
+	@RequestMapping(value = "api/releases/{id}/", method = RequestMethod.GET, produces = "application/json")
+	public ResponseEntity<Release> getRelease(@PathVariable("id") int id) {
 
-        // Rückgabe: Release-Entity mit Liste von Tracks und Master-ID
-        ReleaseService service = new ReleaseService();
-        Release release = service.getReleaseById(id);
+		// Rückgabe: Release-Entity mit Liste von Tracks und Master-ID
+		ReleaseService service = new ReleaseService();
+		Release release = service.getReleaseById(id);
 
-        return new ResponseEntity<Release>(release, HttpStatus.OK);
-    }
+		return new ResponseEntity<Release>(release, HttpStatus.OK);
+	}
 
-    @RequestMapping(value = "api/release/", method = RequestMethod.POST, consumes = "application/json", produces = "application/json")
-    public ResponseEntity<Long> postNewRelease(@RequestBody ReleaseDto release) {
+	@RequestMapping(value = "api/release/", method = RequestMethod.POST, consumes = "application/json", produces = "application/json")
+	public ResponseEntity<Long> postNewRelease(@RequestBody ReleaseDto release) {
 
-        LOG.debug(release.toString());
-        ReleaseService service = new ReleaseService();
-        long id = service.insertRelease(release);
+		LOG.debug(release.toString());
+		ReleaseService service = new ReleaseService();
+		long id = service.insertRelease(release);
 
-        return new ResponseEntity<Long>(id, HttpStatus.OK);
-    }
-    
-    @RequestMapping(value = "api/release/", method = RequestMethod.GET, produces = "application/json")
-    public ResponseEntity<List<Release>> getReleasesByLabelId(@RequestParam(value = "labelId") long labelId) {
-    	
-    	ReleaseService service = new ReleaseService();
-    	
-    	List<Release> releases = service.getReleasesByLabelId(labelId);
-    	
-    	return new ResponseEntity<List<Release>>(releases, HttpStatus.OK);
-    }
+		return new ResponseEntity<Long>(id, HttpStatus.OK);
+	}
+
+	@RequestMapping(value = "api/release/", method = RequestMethod.GET, produces = "application/json")
+	public ResponseEntity<List<Release>> getReleasesByLabelId(
+			@RequestParam(value = "labelId") int labelId) {
+
+		ReleaseService service = new ReleaseService();
+
+		List<Release> releases = service.getReleasesByLabelId(labelId);
+
+		return new ResponseEntity<List<Release>>(releases, HttpStatus.OK);
+	}
 
 }
